requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.2.0
psycopg2-binary==2.9.4
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
